Refactor artifact extraction into separate file and improve chat session handling
- Moved `extractArtifacts` and related functions to new `chat_artifact.go`
- Improved chat session handler with idempotent creation and better error handling
- Added default values for new chat sessions
- Fixed 404 response format for missing sessions
